+++ internal/opaccount/desktop_account_other.go
@@ -1,0 +1,7 @@
+//go:build !darwin || !cgo
+
+package opaccount
+
+func DetectDefaultDesktopAccount() string {
+	return ""
+}